AppleTV Support
There was a problem with the new implementation of TextInputDialog and
AlertDialog that is now fixed.
Also the MFIController plugin now should compile correctly on both
platforms (iOS and tvOS)
I modified the template to use the same library names without the “_tv”
suffix. Also it will preserve the name of the app during the
exportation.