Updated LuaJIT binaries to the new patched version
The new patched version has an extra call so that you can use fake
threads.

The new build-android-ab.sh file is for me!  It's a script for making
Android LuaJIT on 32bit linux.  You need a copy of 32bit version of the
NDK.

The prebuilt_plus.zip file is prebuild versions of LuaJIT for Android
and Win32.  There is also a patch for making Gideros export LuaJIT
bytecode rather than plain Lua bytecode.