lua and luac macro fixes and new "include" directive
Each lua file has it's own `MACRO` table which is cleared at the
beginning of parsing.
To support import of macros from other files "include" directive can be
used as `include "filename"` to load content of the file at the
compilation step, for example:
`include "macro_functions"`
will try to open the file "macro_functions", load it's content and paste
it at the place of `include` directive. Of course, any content of any
file can be included this way, just like in C preprocessor.