change position of find/replace dialog box
Currently the find/replace dialog box is docked to the right of Gideros Studio debug output box (where I usually put Gideros Player!).
This is in "my" case an inconvenient place to be!?

This commit puts the find/replace dialog box in the center of Gideros Studio debug output box. It is up to you do decide ;-)

Compiled and tested on my window 10 machine and it is imho more convenient.