win32 issue with sticky modifier keys
My issue was in win32 when a key modifier was being pressed (SHIFT, ALT, CTRL) it wouldn't get released when no longer pressed (it was like stuck in the pressed state). To cancel the pressed state one had to press the modifier key again!

I tested changing from GetKeyState to GetAsyncKeyState and it seems to have solve the issue.