sticky modifier keys
My issue was in win32 when a key modifier was being pressed (SHIFT, ALT, CTRL) it wouldn't get released when no longer pressed (it was like stuck in the pressed state). To cancel the pressed state one had to press the modifier key again!

I change according to your recommandation and now using if (GetKeyState(VK_xxx) & 0x8000)

I also added & 0x8000 to int getKeyboardModifiers() but kept the GetAsyncKeyState otherwise the modifiers wouldn't work with the mouse.